Collection of fixes and improvements to isolation between fake network instances.

+++ orchestrator/env/shutdown_all.sh
@@ -6,6 +6,12 @@ ips=($(awk '/'$OSM_SUBNET'/ {print $3}' "$HOSTS_DIR/$VOSM_HOSTBR.leasefile"))
 
 for ip in ${ips[@]}
 do
-  echo "Closing down $ip"
-  ssh root@$ip poweroff
+  ping -c1 -W1 $ip >/dev/null
+  if [ "$?" == "0" ]
+  then
+    echo "Closing down $ip"
+    ssh root@$ip poweroff
+  fi
 done
+
+./net_ctrl.sh close $VOSM_HOSTBR $HOSTS_DIR $OSM_SUBNET $OSM_DOMAIN
